# openapi.yml
openapi: 3.0.3
info:
  title: EE3070 Project Server API
  version: 1.0.0
  description: |
    OpenAPI 規格依據目前 `src/routes` 的實際實作整理。
    主要功能包含課程、檔案上傳、題目產生任務、取消與狀態串流 (SSE)。
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /course/{courseId}/question:
    get:
      tags: [ Questions ]
      summary: List question metadata in a course
      operationId: listCourseQuestions
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
      responses:
        '200':
          description: Question metadata list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [ Questions ]
      summary: Create question generation task
      operationId: createQuestionTask
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
            example:
              clientTaskId: e0eb839a-b682-44e5-8e06-85a4a5ef16f1
              prompt: Please generate any 4 understanding-check questions.
      responses:
        '201':
          description: New question generation task enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionCreateResponse'
        '202':
          description: Idempotency hit; existing task id returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionCreateDuplicateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /course/{courseId}/question/{questionId}:
    get:
      tags: [ Questions ]
      summary: Get question result or current status
      operationId: getQuestion
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
        - $ref: '#/components/parameters/QuestionIdPath'
      responses:
        '200':
          description: Question generation finished (DONE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDoneResponse'
        '202':
          description: Question not ready yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionNotReadyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [ Questions ]
      summary: Update question visibility (public/private)
      operationId: updateQuestionVisibility
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
        - $ref: '#/components/parameters/QuestionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionVisibilityUpdateRequest'
            example:
              visibility: public
      responses:
        '200':
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionVisibilityUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Question status is not DONE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionStatusConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ Questions ]
      summary: Delete question
      operationId: deleteQuestion
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
        - $ref: '#/components/parameters/QuestionIdPath'
      responses:
        '200':
          description: Question deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDeleteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Question is still processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionStatusConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /course/{courseId}/question/{questionId}/cancel:
    post:
      tags: [ Questions ]
      summary: Cancel question generation
      operationId: cancelQuestion
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
        - $ref: '#/components/parameters/QuestionIdPath'
      responses:
        '202':
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionCancelAcceptedResponse'
        '200':
          description: |
            Current implementation returns HTTP 200 with `code: 410`
            when question is already in terminal status (DONE/ERROR/CANCELLED/STALE).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionCancelTerminalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /course/{courseId}/question/{questionId}/stream:
    get:
      tags: [ Questions ]
      summary: Stream question status by SSE
      description: |
        Emits `event: status` with payload:
        `{"questionId":"...","status":"PENDING|GENERATING|DONE|ERROR|CANCELLED|STALE"}`.
        When status is DONE and title exists, it also emits `event: delta` with
        `{"questionId":"...","title":"..."}`.
      operationId: streamQuestionStatus
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
        - $ref: '#/components/parameters/QuestionIdPath'
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: status
                data: {"questionId":"829c0cf0-791c-43f4-9e55-dc4d27f68d63","status":"PENDING"}

                event: status
                data: {"questionId":"829c0cf0-791c-43f4-9e55-dc4d27f68d63","status":"GENERATING"}

                event: delta
                data: {"questionId":"829c0cf0-791c-43f4-9e55-dc4d27f68d63","title":"Understanding Java Exceptions"}

                event: status
                data: {"questionId":"829c0cf0-791c-43f4-9e55-dc4d27f68d63","status":"DONE"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    CourseIdPath:
      name: courseId
      in: path
      required: true
      description: Course UUID
      schema:
        type: string
        format: uuid
...
    QuestionIdPath:
      name: questionId
      in: path
      required: true
      description: Question UUID
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: Invalid request
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            message: Resource not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: Internal server error
  schemas:
    ErrorResponse:
      type: object
      required: [ code, message ]
      properties:
        code:
          type: integer
        message:
          type: string
...
    QuestionStatus:
      type: string
      enum: [ PENDING, GENERATING, DONE, ERROR, CANCELLED, STALE ]
    QuestionMeta:
      type: object
      properties:
        visibility:
          oneOf:
            - type: string
              enum: [ '0', '1', 'public', 'private' ]
            - type: integer
              enum: [ 0, 1 ]
        createAt:
          type: string
          description: ISO-8601 string or empty string
        finishedAt:
          type: string
          description: ISO-8601 string or empty string
        errorMessage:
          type: string
        questionId:
          type: string
          format: uuid
        updateAt:
          type: string
        startAt:
          type: string
          description: ISO-8601 string or empty string
        title:
          type: string
        prompt:
          type: string
        status:
          $ref: '#/components/schemas/QuestionStatus'
        courseId:
          type: string
          format: uuid
      required:
        - questionId
        - courseId
        - status
    QuestionListResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: All question meta get successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuestionMeta'
    QuestionCreateRequest:
      type: object
      required: [ clientTaskId ]
      properties:
        clientTaskId:
          type: string
          format: uuid
        prompt:
          type: string
          nullable: true
    QuestionCreateResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 201
        message:
          type: string
          example: Question generate request enqueued.
        data:
          type: object
          required: [ questionId, title, status ]
          properties:
            questionId:
              type: string
              format: uuid
            title:
              type: string
            status:
              type: string
              enum: [ PENDING ]
    QuestionCreateDuplicateResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 202
        message:
          type: string
          example: Request already receive
        data:
          type: object
          required: [ questionId ]
          properties:
            questionId:
              type: string
              format: uuid
    QuestionDoneResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          type: object
          required: [ meta, question, mark ]
          properties:
            meta:
              $ref: '#/components/schemas/QuestionMeta'
            question:
              type: array
              nullable: true
              items:
                type: string
            mark:
              type: object
              nullable: true
              additionalProperties: true
    QuestionNotReadyResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 202
        message:
          type: string
        data:
          type: object
          required: [ meta ]
          properties:
            meta:
              $ref: '#/components/schemas/QuestionMeta'
    QuestionVisibilityUpdateRequest:
      type: object
      required: [ visibility ]
      properties:
        visibility:
          type: string
          enum: [ public, private ]
    QuestionVisibilityUpdateResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          type: object
          required: [ visibility ]
          properties:
            visibility:
              type: string
              enum: [ public, private ]
    QuestionStatusConflictResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 409
        message:
          type: string
        data:
          type: object
          required: [ status ]
          properties:
            status:
              $ref: '#/components/schemas/QuestionStatus'
    QuestionDeleteResponse:
      type: object
      required: [ code, message ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
    QuestionCancelAcceptedResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 202
        message:
          type: string
        data:
          type: object
          required: [ status ]
          properties:
            status:
              type: string
              enum: [ CANCELLED ]
    QuestionCancelTerminalResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 410
        message:
          type: string
        data:
          type: object
          required: [ status ]
          properties:
            status:
              $ref: '#/components/schemas/QuestionStatus'